PR feedback
mostly moving things around
