+++ crates/forge/tests/cli/zk_build.rs
@@ -1,4 +1,3 @@
-use foundry_common::fs;
 use foundry_test_utils::util::OutputExt;
 use regex::Regex;
 
@@ -11,68 +10,4 @@ forgetest_init!(test_zk_build_sizes, |prj, cmd| {
             .unwrap();
 
     assert!(pattern.is_match(&stdout), "Unexpected size output:\n{stdout}");
-});
-
-// tests cache works as expected in zksync mode
-forgetest_init!(test_zk_cache_ok, |prj, cmd| {
-    let zk_toml = r#"[profile.default]
-src = 'src'
-out = 'out'
-libs = ['lib']
-solc = '0.8.26'
-
-[profile.default.zksync]
-zksolc = '1.5.6'
-"#;
-
-    fs::write(prj.root().join("foundry.toml"), zk_toml).unwrap();
-
-    cmd.args(["build", "--zksync"]);
-    let stdout_1 = cmd.assert_success().get_output().stdout_lossy();
-    let pattern_1 = Regex::new(r"Compiler run successful").unwrap();
-
-    let stdout_2 = cmd.assert_success().get_output().stdout_lossy();
-    let pattern_2 = Regex::new(r"No files changed, compilation skipped").unwrap();
-
-    assert!(pattern_1.is_match(&stdout_1));
-    assert!(pattern_2.is_match(&stdout_2));
-});
-
-// tests cache is invalidated when zksolc version changes
-forgetest_init!(test_zk_cache_invalid_on_version_changed, |prj, cmd| {
-    let template_toml = r#"[profile.default]
-src = 'src'
-out = 'out'
-libs = ['lib']
-solc = '0.8.26'
-
-[profile.default.zksync]
-"#;
-
-    let toml_156 = format!(
-        r#"{template_toml}
-zksolc = '1.5.6'
-"#
-    );
-
-    let toml_157 = format!(
-        r#"{template_toml}
-zksolc = '1.5.7'
-"#
-    );
-
-    fs::write(prj.root().join("foundry.toml"), toml_156).unwrap();
-
-    cmd.args(["build", "--zksync"]);
-    let stdout_1 = cmd.assert_success().get_output().stdout_lossy();
-    let pattern_1 = Regex::new(r"Compiler run successful").unwrap();
-
-    fs::remove_file(prj.root().join("foundry.toml")).unwrap();
-    fs::write(prj.root().join("foundry.toml"), toml_157).unwrap();
-
-    let stdout_2 = cmd.assert_success().get_output().stdout_lossy();
-    let pattern_2 = Regex::new(r"Compiler run successful!").unwrap(); // if we see this, means the cache was invalidated
-
-    assert!(pattern_1.is_match(&stdout_1));
-    assert!(pattern_2.is_match(&stdout_2));
 });
