+++ crates/zksync/compilers/tests/zksync_tests.rs
@@ -12,6 +12,7 @@ use foundry_test_utils::foundry_compilers::{
 };
 
 use era_solc::standard_json::input::settings::{error_type::ErrorType, warning_type::WarningType};
+use foundry_test_utils::{forgetest_init, util::OutputExt};
 use foundry_zksync_compilers::{
     artifacts::{contract::Contract, error::Error},
     compilers::{
@@ -19,6 +20,7 @@ use foundry_zksync_compilers::{
         zksolc::{input::ZkSolcInput, ZkSolc, ZkSolcCompiler, ZkSolcSettings},
     },
 };
+use regex::Regex;
 
 #[test]
 fn zksync_can_compile_dapp_sample() {
@@ -654,3 +656,69 @@ fn zksync_can_compile_yul_sample() {
 
     assert!(!yul_bytecode.is_empty(), "SimpleStore bytecode is empty");
 }
+
+forgetest_init!(test_zk_cache_ok, |prj, cmd| {
+    let zk_toml = r#"[profile.default]
+src = 'src'
+out = 'out'
+libs = ['lib']
+solc = '0.8.26'
+
+[profile.default.zksync]
+zksolc = '1.5.6'
+"#;
+
+    fs::write(prj.root().join("foundry.toml"), zk_toml).unwrap();
+
+    cmd.args(["build", "--zksync"]);
+    let stdout_1 = cmd.assert_success().get_output().stdout_lossy();
+    let pattern_1 = Regex::new(r"Compiler run successful").unwrap();
+
+    let stdout_2 = cmd.assert_success().get_output().stdout_lossy();
+    let pattern_2 = Regex::new(r"No files changed, compilation skipped").unwrap();
+
+    assert!(pattern_1.is_match(&stdout_1));
+    assert!(pattern_2.is_match(&stdout_2));
+});
+
+// tests cache is invalidated when zksolc version changes
+forgetest_init!(test_zk_cache_invalid_on_version_changed, |prj, cmd| {
+    let template_toml = r#"[profile.default]
+src = 'src'
+out = 'out'
+libs = ['lib']
+solc = '0.8.26'
+
+[profile.default.zksync]
+"#;
+
+    let toml_156 = format!(
+        r#"{template_toml}
+zksolc = '1.5.6'
+"#
+    );
+
+    let toml_157 = format!(
+        r#"{template_toml}
+zksolc = '1.5.7'
+"#
+    );
+
+    fs::write(prj.root().join("foundry.toml"), toml_156).unwrap();
+
+    cmd.args(["build", "--zksync"]);
+    let stdout_1 = cmd.assert_success().get_output().stdout_lossy();
+    let pattern_1 = Regex::new(r"Compiler run successful").unwrap();
+
+    fs::remove_file(prj.root().join("foundry.toml")).unwrap();
+    fs::write(prj.root().join("foundry.toml"), toml_157).unwrap();
+
+    let stdout_2 = cmd.assert_success().get_output().stdout_lossy();
+    let pattern_2 = Regex::new(r"Compiler run successful!").unwrap(); // if we see this, means the cache was invalidated
+
+    print!("stdout_1: {}", stdout_1);
+    print!("stdout_2: {}", stdout_2);
+
+    assert!(pattern_1.is_match(&stdout_1));
+    assert!(pattern_2.is_match(&stdout_2));
+});
